packages/qfull-nuclear/pyproject.toml
# qfull-nuclear: ab-initio nuclear domain plugin for QCompass.

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "qfull-nuclear"
version = "0.1.0"
description = "QFull-Nuclear — ab-initio nuclear structure / 0νββ toy / NCSM domain plugin for QCompass."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "QCompass authors" }]
dependencies = [
  "pydantic>=2.12",
  "numpy>=1.26",
  "scipy>=1.14",
  "h5py>=3.12",
  "pyyaml>=6",
  "qcompass-core>=0.1,<0.2",
]

[project.optional-dependencies]
classical = ["block2>=0.5.3"]
ionq = ["amazon-braket-sdk>=1.85"]
ibm = ["qiskit>=2.2,<3", "qiskit-aer>=0.17", "qiskit-ibm-runtime"]
shellmodel = []  # pyNSMC / nushellx — research, vendor on demand
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.24",
  "pytest-regressions>=2.5",
  "hypothesis>=6.120",
  "mypy>=1.13",
  "ruff>=0.8",
]

[project.entry-points."qcompass.domains"]
nuclear = "qfull_nuc.sim:NuclearSimulation"

[tool.hatch.build.targets.wheel]
packages = ["src/qfull_nuc"]

[tool.hatch.build.targets.wheel.force-include]
"src/qfull_nuc/instances" = "qfull_nuc/instances"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true
plugins = ["pydantic.mypy"]
disallow_untyped_defs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
  "block2.*", "pyblock2.*",
  "qiskit.*", "qiskit_aer.*", "qiskit_ibm_runtime.*",
  "braket.*",
  "pynsmc.*", "nushellx.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests", "src/qfull_nuc/audit"]
asyncio_mode = "auto"
markers = [
  "slow: long-running test",
  "s_audit: per-domain S-nuc-* audit",
  "needs_block2: requires pyblock2",
  "needs_qiskit: requires qiskit",
]
